MAJOR refactoring of panels block processor. Update view presentation of docs panels
Changed the width attribute to scale in the adoc files to better control the size of the documents. Simplified PanelsBlockProcessor in the Kotlin files for better readability and maintainability. Updated the HTML files to better present the image blocks. These changes were made for better UI and improved code structure.