Update default values in DocOpsBlockProcessor and document usage
Fix the default value for title attribute in AbstractDocOpsBlockProcessor and updated image URL in RoadmapBlockProcessor. Ensured that the title of the roadmap can now be customized. Additionally, updated the documentation on how to create a timeline and what each of the macro options does. The improvement was necessary to provide user flexibility in title attribute and to better instruct users on manual timeline creation, thus enhancing user experience.