Update SVG generation methods
Updated the methods for SVG generation in 'AbstractDocOpsBlockProcessor' and 'TimelineBlockProcessor' to include all lines in the SVG URL and to standardise the SVG format regardless of the structure type. This change was necessary to fix an issue where only the first line of the SVG URL was being output, leading to incomplete and incorrect SVGs. Also, the HTML files were changed to reflect the improvements in SVGs.