Add iconic.css file
The iconic.css file has been added to the project. This file is important for loading external fonts and setting general styles such as colors, alignments, and sizes, which will help in keep consistency across the application. It also provides necessary resets for tables, lists, and typography. This file will serve as the primary css file for the project, and its settings will be inherited by other css files.