Optimize code in PanelsBlockProcessor for shorter syntax.
Update the code in PanelsBlockProcessor.kt and several other files to consolidate the options generation into a single place. It was previously scattered across these files and it was harder to change in future if the logic changes. This makes the overfall file shorter and easier to read. Now the options is generated in one method in AbstractDocOpsBlockProcessor.kt and there is less redundancy across the files. The update also includes corresponding changes in HTML.