Add support for buttons block for Docs Panel
Added a new Kotlin class for processing buttons block. Registered this new class in the panels registry.
Made corresponding changes in the panel HTML. Added buttons block listing in the sidebar navigation. Made minor changes in the image block and added SVG for the newly created button block. The goal of this commit is to extend functionality for Docs Panel and allow buttons interface in addition to previously existing options.