Add dark mode option in panel extension
This change allows users to specify a dark mode theme in the DocOps AsciiDoctorJ panel extension.
A new boolean parameter "useDark" is added to the relevant methods, which indicates whether to use the dark mode or not.
The default value of "useDark" is false, preserving the current behavior if not specified by the user.
This option is controlled by users via the "useDark" attribute. If "useDark" is set to "true", the extension operates in the dark mode.
This enhances user experience as it provides an alternative theme that is easier on the eyes under low-light conditions.